+++ src/components/Torrent/DashboardItems/Availability.vue
@@ -1,7 +1,7 @@
 <template>
   <v-flex xs6 sm1 md1>
     <div class="caption grey--text">
-      {{ $t('torrent.availability') | titleCase }}
+      {{ $t('torrent.properties.availability') | titleCase }}
     </div>
     <div>
       {{ availability }}