+++ src/components/Torrent/DashboardItems/CompletedOn.vue
@@ -1,7 +1,7 @@
 <template>
   <v-flex xs6 sm2>
     <div class="caption grey--text">
-      {{ $t('torrent.completed_on') | titleCase }}
+      {{ $t('torrent.properties.completed_on') | titleCase }}
     </div>
     <div>{{ torrent.completed_on }}</div>
   </v-flex>