+++ src/components/Torrent/DashboardItems/Peers.vue
@@ -1,11 +1,11 @@
 <template>
   <v-flex xs6 sm1 md1>
     <div class="caption grey--text">
-      {{ $t('peers') }}
+      {{ $t('torrent.properties.peers') }}
     </div>
     <div>
       {{ torrent.num_leechs }}
-      <span class="grey--text caption"> /{{ torrent.available_peers }} </span>
+      <span class="grey--text caption"> / {{ torrent.available_peers }} </span>
     </div>
   </v-flex>
 </template>
@@ -15,7 +15,7 @@ import { defineComponent } from 'vue'
 import { Torrent } from '@/models'
 
 export default defineComponent({
-  name: 'DashboardPeers',
+  name: 'Peers',
   props: {
     torrent: Torrent
   }