+++ src/components/Torrent/DashboardItems/Ratio.vue
@@ -1,7 +1,7 @@
 <template>
   <v-flex xs6 sm1 md1>
     <div class="caption grey--text">
-      {{ $t('ratio') }}
+      {{ $t('torrent.properties.ratio') }}
     </div>
     <div>{{ torrent.ratio }}</div>
   </v-flex>