+++ src/components/Torrent/DashboardItems/SavePath.vue
@@ -1,11 +1,10 @@
 <template>
-  <v-flex xs6 sm1 md1>
+  <v-flex xs6 md2 sm1>
     <div class="caption grey--text">
-      {{ $t('torrent.properties.global_speed') }}
+      {{ $t('torrent.properties.save_path') | titleCase }}
     </div>
-    <div>
-      {{ torrent.globalSpeed | getDataValue(1) }}
-      <span class="caption grey--text"> {{ torrent.globalSpeed | getDataUnit }}/s </span>
+    <div class="truncate" :title="torrent.savePath">
+      {{ torrent.savePath }}
     </div>
   </v-flex>
 </template>
@@ -15,7 +14,7 @@ import { defineComponent } from 'vue'
 import { Torrent } from '@/models'
 
 export default defineComponent({
-  name: 'GlobalSpeed',
+  name: 'SavePath',
   props: {
     torrent: Torrent
   }