+++ src/components/Torrent/DashboardItems/SeenComplete.vue
@@ -1,12 +1,9 @@
 <template>
   <v-flex xs6 sm1 md1>
     <div class="caption grey--text">
-      {{ $t('torrent.properties.global_speed') }}
+      {{ $t('torrent.properties.seen_complete') | titleCase }}
     </div>
-    <div>
-      {{ torrent.globalSpeed | getDataValue(1) }}
-      <span class="caption grey--text"> {{ torrent.globalSpeed | getDataUnit }}/s </span>
-    </div>
+    <div>{{ torrent.seen_complete }}</div>
   </v-flex>
 </template>
 
@@ -15,7 +12,7 @@ import { defineComponent } from 'vue'
 import { Torrent } from '@/models'
 
 export default defineComponent({
-  name: 'GlobalSpeed',
+  name: 'SeenComplete',
   props: {
     torrent: Torrent
   }